requirements.txt
torch==2.2.1
torchvision==0.17.1
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.3
albumentations==1.4.2
pillow==10.2.0
pandas==2.2.1
